Switching to stable version of rascal

+++ Jenkinsfile
@@ -27,10 +27,12 @@ node {
     }
 
     stage('Deploy update site') {
-	     	sh 'cd typhonql-update-site && mvn clean package'
-		sh "rm -rf ${UPDATE_SITE_PATH}"
-		sh "mkdir ${UPDATE_SITE_PATH}"
-        sh "cp -a typhonql-update-site/target/. ${UPDATE_SITE_PATH}/"
+        if (env.BRANCH_NAME == "master") {
+            sh 'cd typhonql-update-site && mvn clean package'
+            sh "rm -rf ${UPDATE_SITE_PATH}"
+            sh "mkdir ${UPDATE_SITE_PATH}"
+            sh "cp -a typhonql-update-site/target/. ${UPDATE_SITE_PATH}/"
+        }
     }
     }catch (e){
         currentBuild.result = "FAILED"
